+++ src/main/java/io/github/alexengrig/lambdax/function/PredicateB.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Alexengrig Dev.
+ * Copyright 2019 - 2020 Alexengrig Dev.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ import java.util.function.Predicate;
  * @param <T> the type of the input to the predicate
  * @param <R> the type of the mapper result
  * @author Grig Alex
- * @version 0.2.0
+ * @version 0.2.1
  * @see java.util.Comparator
  * @see java.util.Objects
  * @see java.util.function.Function
@@ -36,14 +36,6 @@ import java.util.function.Predicate;
  * @since 0.2.0
  */
 /* package */class PredicateB<T, R> implements PredicateI<T, R> {
-    /**
-     * <p>The value for compare with the comparator result</p>
-     *
-     * @see java.util.Comparator#compare(Object, Object)
-     * @since 0.2.0
-     */
-    private final static int ZERO = 0;
-
     /**
      * <p>The mapper of the input to the predicate.</p>
      *
@@ -65,13 +57,14 @@ import java.util.function.Predicate;
 
     /**
      * <p>
-     * Returns the {@link io.github.alexengrig.lambdax.function.PredicateB} with the new mapper for
-     * the input to the predicate.
+     * Returns the {@link io.github.alexengrig.lambdax.function.PredicateB} with
+     * the new mapper for the input to the predicate.
      * </p>
      *
      * @param mapper a function of map the input
      * @param <V>    a new type of the mapper result
-     * @return The {@link io.github.alexengrig.lambdax.function.PredicateI} with the new mapper
+     * @return The {@link io.github.alexengrig.lambdax.function.PredicateI} with
+     * the new mapper
      * @see io.github.alexengrig.lambdax.function.PredicateI
      * @see java.util.function.Function
      * @see io.github.alexengrig.lambdax.function.PredicateB
@@ -85,13 +78,16 @@ import java.util.function.Predicate;
 
     /**
      * <p>
-     * Returns the {@link io.github.alexengrig.lambdax.function.ComparablePredicateB} with the new mapper with
-     * comparable result for the input to the predicate.
+     * Returns the {@link
+     * io.github.alexengrig.lambdax.function.ComparablePredicateB} with the new
+     * mapper with comparable result for the input to the predicate.
      * </p>
      *
      * @param mapper a function of map the input to the comparable result
      * @param <V>    a new comparable type of the mapper result
-     * @return The {@link io.github.alexengrig.lambdax.function.ComparablePredicateI} with the new mapper
+     * @return The {@link
+     * io.github.alexengrig.lambdax.function.ComparablePredicateI} with the
+     * new mapper
      * @see java.lang.Comparable
      * @see io.github.alexengrig.lambdax.function.ComparablePredicateI
      * @see io.github.alexengrig.lambdax.function.ComparableResultFunction
@@ -105,8 +101,63 @@ import java.util.function.Predicate;
     }
 
     /**
+     * {@inheritDoc}
+     *
+     * @since 0.2.1
+     */
+    @Override
+    public <V> OptionalPredicateI<T, V> mapToNullable(Function<R, V> mapper) {
+        return new OptionalPredicateB<>(function.andThen(mapper));
+    }
+
+    /**
+     * {@inheritDoc}
+     * @since 0.2.1
+     */
+    @Override
+    public <V extends Comparable<V>> ComparableOptionalPredicateI<T, V> mapToNullable(
+            ComparableResultFunction<R, V> mapper) {
+        return new ComparableOptionalPredicateB<>(function.andThen(mapper));
+    }
+
+    /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result via the checker.
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result is {@code null}.
+     * </p>
+     *
+     * @return The {@link java.util.function.Predicate} with compare
+     * @see java.util.function.Predicate
+     * @see java.util.Objects#isNull(Object)
+     * @see java.util.function.Function#apply(Object)
+     * @since 0.2.0
+     */
+    @Override
+    public Predicate<T> isNull() {
+        return t -> Objects.isNull(function.apply(t));
+    }
+
+    /**
+     * <p>
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result is not {@code null}.
+     * </p>
+     *
+     * @return The {@link java.util.function.Predicate} with compare
+     * @see java.util.function.Predicate
+     * @see java.util.Objects#nonNull(Object)
+     * @see java.util.function.Function#apply(Object)
+     * @since 0.2.0
+     */
+    @Override
+    public Predicate<T> nonNull() {
+        return t -> Objects.nonNull(function.apply(t));
+    }
+
+    /**
+     * <p>
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result via the checker.
      * </p>
      *
      * @param checker a predicate for check the mapper result
@@ -123,39 +174,8 @@ import java.util.function.Predicate;
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result is {@code null}.
-     * </p>
-     *
-     * @return The {@link java.util.function.Predicate} with compare
-     * @see java.util.function.Predicate
-     * @see java.util.Objects#isNull(Object)
-     * @see java.util.function.Function#apply(Object)
-     * @since 0.2.0
-     */
-    @Override
-    public Predicate<T> isNull() {
-        return t -> Objects.isNull(function.apply(t));
-    }
-
-    /**
-     * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result is not {@code null}.
-     * </p>
-     *
-     * @return The {@link java.util.function.Predicate} with compare
-     * @see java.util.function.Predicate
-     * @see java.util.Objects#nonNull(Object)
-     * @see java.util.function.Function#apply(Object)
-     * @since 0.2.0
-     */
-    @Override
-    public Predicate<T> nonNull() {
-        return t -> Objects.nonNull(function.apply(t));
-    }
-
-    /**
-     * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is equals to the other object.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is equals to the other object.
      * </p>
      *
      * @param other an object for compare
@@ -172,12 +192,13 @@ import java.util.function.Predicate;
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is less than the other object
-     * via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is less than the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
@@ -187,17 +208,18 @@ import java.util.function.Predicate;
      */
     @Override
     public Predicate<T> less(R other, Comparator<R> comparator) {
-        return t -> comparator.compare(function.apply(t), other) < ZERO;
+        return t -> Objects.compare(function.apply(t), other, comparator) < 0;
     }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is greater than the other object
-     * via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is greater than the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
@@ -207,6 +229,24 @@ import java.util.function.Predicate;
      */
     @Override
     public Predicate<T> greater(R other, Comparator<R> comparator) {
-        return t -> comparator.compare(function.apply(t), other) > ZERO;
+        return t -> Objects.compare(function.apply(t), other, comparator) > 0;
+    }
+
+    /**
+     * {@inheritDoc}
+     * @since 0.2.1
+     */
+    @Override
+    public Predicate<T> lessOrEqual(R other, Comparator<R> comparator) {
+        return t -> Objects.compare(function.apply(t), other, comparator) <= 0;
+    }
+
+    /**
+     * {@inheritDoc}
+     * @since 0.2.1
+     */
+    @Override
+    public Predicate<T> greaterOrEqual(R other, Comparator<R> comparator) {
+        return t -> Objects.compare(function.apply(t), other, comparator) >= 0;
     }
 }