+++ src/main/java/io/github/alexengrig/lambdax/function/PredicateI.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Alexengrig Dev.
+ * Copyright 2019 - 2020 Alexengrig Dev.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package io.github.alexengrig.lambdax.function;
 
 import java.util.Comparator;
+import java.util.Objects;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
@@ -26,7 +27,7 @@ import java.util.function.Predicate;
  * @param <T> the type of the input to the predicate
  * @param <R> the type of the mapper result
  * @author Grig Alex
- * @version 0.2.0
+ * @version 0.2.1
  * @see java.lang.Comparable
  * @see java.util.Comparator
  * @see java.util.function.Function
@@ -38,13 +39,14 @@ import java.util.function.Predicate;
 public interface PredicateI<T, R> {
     /**
      * <p>
-     * Returns the {@link io.github.alexengrig.lambdax.function.PredicateI} with the new mapper for
-     * the input to the predicate.
+     * Returns the {@link io.github.alexengrig.lambdax.function.PredicateI} with
+     * the new mapper for the input to the predicate.
      * </p>
      *
      * @param mapper a function of map the input
      * @param <V>    a new type of the mapper result
-     * @return The {@link io.github.alexengrig.lambdax.function.PredicateI} with the new mapper
+     * @return The {@link io.github.alexengrig.lambdax.function.PredicateI} with
+     * the new mapper
      * @see io.github.alexengrig.lambdax.function.PredicateI
      * @see java.util.function.Function
      * @since 0.2.0
@@ -53,13 +55,16 @@ public interface PredicateI<T, R> {
 
     /**
      * <p>
-     * Returns the {@link io.github.alexengrig.lambdax.function.ComparablePredicateI} with the new mapper with
-     * comparable result for the input to the predicate.
+     * Returns the {@link
+     * io.github.alexengrig.lambdax.function.ComparablePredicateI} with the new
+     * mapper with comparable result for the input to the predicate.
      * </p>
      *
      * @param mapper a function of map the input to the comparable result
      * @param <V>    a new comparable type of the mapper result
-     * @return The {@link io.github.alexengrig.lambdax.function.ComparablePredicateI} with the new mapper
+     * @return The {@link
+     * io.github.alexengrig.lambdax.function.ComparablePredicateI} with the
+     * new mapper
      * @see java.lang.Comparable
      * @see io.github.alexengrig.lambdax.function.ComparablePredicateI
      * @see io.github.alexengrig.lambdax.function.ComparableResultFunction
@@ -69,7 +74,40 @@ public interface PredicateI<T, R> {
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result via the checker.
+     * Returns the {@link io.github.alexengrig.lambdax.function.OptionalPredicateI} with
+     * the new mapper for the input to the predicate.
+     * </p>
+     *
+     * @param mapper a function of map the input
+     * @param <V>    a new type of the mapper result
+     * @return The {@link io.github.alexengrig.lambdax.function.OptionalPredicateI} with the new mapper
+     * @see io.github.alexengrig.lambdax.function.OptionalPredicateI
+     * @see java.util.function.Function
+     * @since 0.2.1
+     */
+    <V> OptionalPredicateI<T, V> mapToNullable(Function<R, V> mapper);
+
+    /**
+     * <p>
+     * Returns the {@link
+     * io.github.alexengrig.lambdax.function.ComparableOptionalPredicateI} with the new
+     * mapper with comparable result for the input to the predicate.
+     * </p>
+     *
+     * @param mapper a function of map the input to the comparable result
+     * @param <V>    a new comparable type of the mapper result
+     * @return The {@link io.github.alexengrig.lambdax.function.ComparableOptionalPredicateI} with the new mapper
+     * @see java.lang.Comparable
+     * @see io.github.alexengrig.lambdax.function.ComparableOptionalPredicateI
+     * @see io.github.alexengrig.lambdax.function.ComparableResultFunction
+     * @since 0.2.1
+     */
+    <V extends Comparable<V>> ComparableOptionalPredicateI<T, V> mapToNullable(ComparableResultFunction<R, V> mapper);
+
+    /**
+     * <p>
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result via the checker.
      * </p>
      *
      * @param checker a predicate for check the mapper result
@@ -81,29 +119,36 @@ public interface PredicateI<T, R> {
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result is {@code null}.
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result is {@code null}.
      * </p>
      *
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @since 0.2.0
      */
-    Predicate<T> isNull();
+    default Predicate<T> isNull() {
+        return check(Objects::isNull);
+    }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks the mapper result is not {@code null}.
+     * Returns the {@link java.util.function.Predicate} that checks the mapper
+     * result is not {@code null}.
      * </p>
      *
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @since 0.2.0
      */
-    Predicate<T> nonNull();
+    default Predicate<T> nonNull() {
+        return check(Objects::nonNull);
+    }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is equals to the other object.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is equals to the other object.
      * </p>
      *
      * @param other an object for compare
@@ -111,75 +156,87 @@ public interface PredicateI<T, R> {
      * @see java.util.function.Predicate
      * @since 0.2.0
      */
-    Predicate<T> equal(R other);
+    default Predicate<T> equal(R other) {
+        return check(other::equals);
+    }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is less than the other object
-     * via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is less than the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
      * @since 0.2.0
      */
-    Predicate<T> less(R other, Comparator<R> comparator);
+    default Predicate<T> less(R other, Comparator<R> comparator) {
+        return check(r -> Objects.compare(r, other, comparator) < 0);
+    }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is greater than the other object
-     * via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is greater than the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
      * @since 0.2.0
      */
-    Predicate<T> greater(R other, Comparator<R> comparator);
+    default Predicate<T> greater(R other, Comparator<R> comparator) {
+        return check(r -> Objects.compare(r, other, comparator) > 0);
+    }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is less than or
-     * equals to the other object via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is less than or equals to the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
-     * @see io.github.alexengrig.lambdax.function.PredicateI#less(Object, Comparator)
+     * @see io.github.alexengrig.lambdax.function.PredicateI#less(Object,
+     * Comparator)
      * @see java.util.function.Predicate#or(Predicate)
      * @see io.github.alexengrig.lambdax.function.PredicateI#equal(Object)
      * @since 0.2.0
      */
     default Predicate<T> lessOrEqual(R other, Comparator<R> comparator) {
-        return less(other, comparator).or(equal(other));
+        return check(r -> Objects.compare(r, other, comparator) <= 0);
     }
 
     /**
      * <p>
-     * Returns the {@link java.util.function.Predicate} that checks if the input object is greater than or
-     * equals to the other object via the comparator.
+     * Returns the {@link java.util.function.Predicate} that checks if the input
+     * object is greater than or equals to the other object via the comparator.
      * </p>
      *
      * @param other      an object for compare
-     * @param comparator a comparison function for compare the input object with the other object
+     * @param comparator a comparison function for compare the input object with
+     *                   the other object
      * @return The {@link java.util.function.Predicate} with compare
      * @see java.util.function.Predicate
      * @see java.util.Comparator
-     * @see io.github.alexengrig.lambdax.function.PredicateI#greater(Object, Comparator)
+     * @see io.github.alexengrig.lambdax.function.PredicateI#greater(Object,
+     * Comparator)
      * @see java.util.function.Predicate#or(Predicate)
      * @see io.github.alexengrig.lambdax.function.PredicateI#equal(Object)
      * @since 0.2.0
      */
     default Predicate<T> greaterOrEqual(R other, Comparator<R> comparator) {
-        return greater(other, comparator).or(equal(other));
+        return check(r -> Objects.compare(r, other, comparator) >= 0);
     }
 }