+++ CHANGES.md
@@ -1,5 +1,11 @@
 # *Lambda*X changes
 
+## 0.2.1
+
+*   [[#86](../../issues/86)] Create PredicateX#ofNullable with null safe mappable predicate
+    *   Create FunctionX#nullSafe for null safe function
+*   [[#92](../../issues/92)|[#89](../../issues/89)] Remove wildcards from collection
+
 ## 0.2.0
 
 Create the function package